Updated 2-do_deploy_web_static.py

+++ 2-do_deploy_web_static.py
@@ -18,6 +18,8 @@ def do_deploy(archive_path):
     try:
         # Upload the archive to the /tmp/ directory of the web server
         put(archive_path, "/tmp/")
+        # copy the archive to the tmp folder in the local machine
+        local("cp {} /tmp/".format(archive_path))
 
         # Uncompress the archive to the folder
         # /data/web_static/releases/<archive filename without extension>
@@ -27,34 +29,59 @@ def do_deploy(archive_path):
         file_name = archive_path.split("/")[-1]
         # get the folder name without the extension
         folder_name = file_name.split(".")[0]
-        run("mkdir -p /data/web_static/releases/{}/".format(folder_name))
-        run("tar -xzf /tmp/{} -C /data/web_static/releases/{}/".format(
-            file_name, folder_name))
 
-        # Delete the archive from the web server
-        run("rm /tmp/{}".format(file_name))
-        # move all the files to parent folder
-        old = "/data/web_static/releases/{}/web_static/*".format(folder_name)
-        new = "/data/web_static/releases/{}/".format(folder_name)
-        run("mv {} {}".format(old, new))
+        def deploy_local(file_name, folder_name):
+            local("mkdir -p /data/web_static/releases/{}/".format(folder_name))
+            local("tar -xzf /tmp/{} -C /data/web_static/releases/{}/".format(
+                file_name, folder_name))
 
-        # run(
-        #     "mv /data/web_static/releases/{}/web_static/*
-        # /data/web_static/releases/{}/"
-        #     .format(folder_name, folder_name))
-        # Delete empty folder after moving content
-        run("rm -rf /data/web_static/releases/{}/web_static".format(
-            folder_name))
+            # Delete the archive from the web server
+            local("rm /tmp/{}".format(file_name))
+            # move all the files to parent folder
+            old = "/data/web_static/releases/{}/web_static/*".format(folder_name)
+            new = "/data/web_static/releases/{}/".format(folder_name)
+            local("mv {} {}".format(old, new))
 
-        # Delete symbolic link /data/web_static/current from the web server
-        run("rm -rf /data/web_static/current")
+            local("rm -rf /data/web_static/releases/{}/web_static".format(
+                folder_name))
 
-        # Create a new the symbolic link /data/web_static/current on
-        # the web server, linked to the new version of your code
-        # (/data/web_static/releases/<archive filename without extension>)
-        run(
-            "ln -s /data/web_static/releases/{}/ /data/web_static/current"
-            .format(folder_name))
+            # Delete symbolic link /data/web_static/current from the web server
+            local("rm -rf /data/web_static/current")
+
+            # Create a new the symbolic link /data/web_static/current on
+            # the web server, linked to the new version of your code
+            # (/data/web_static/releases/<archive filename without extension>)
+            local(
+                "ln -s /data/web_static/releases/{}/ /data/web_static/current"
+                .format(folder_name))
+            return true
+        
+        locallydep = do_deploy_local(file_name, folder_name)   
+ 
+        if locallydep:
+            run("mkdir -p /data/web_static/releases/{}/".format(folder_name))
+            run("tar -xzf /tmp/{} -C /data/web_static/releases/{}/".format(
+                file_name, folder_name))
+
+            # Delete the archive from the web server
+            run("rm /tmp/{}".format(file_name))
+            # move all the files to parent folder
+            old = "/data/web_static/releases/{}/web_static/*".format(folder_name)
+            new = "/data/web_static/releases/{}/".format(folder_name)
+            run("mv {} {}".format(old, new))
+
+            run("rm -rf /data/web_static/releases/{}/web_static".format(
+                folder_name))
+
+            # Delete symbolic link /data/web_static/current from the web server
+            run("rm -rf /data/web_static/current")
+
+            # Create a new the symbolic link /data/web_static/current on
+            # the web server, linked to the new version of your code
+            # (/data/web_static/releases/<archive filename without extension>)
+            run(
+                "ln -s /data/web_static/releases/{}/ /data/web_static/current"
+                .format(folder_name))
 
         return True
